move loss functions to metrics
