+++ chemprop/nn/__init__.py
@@ -6,21 +6,6 @@ from .agg import (
     NormAggregation,
     SumAggregation,
 )
-from .loss import (
-    MSE,
-    SID,
-    BCELoss,
-    BinaryDirichletLoss,
-    BinaryMCCLoss,
-    ChempropMetric,
-    CrossEntropyLoss,
-    EvidentialLoss,
-    LossFunctionRegistry,
-    MulticlassDirichletLoss,
-    MulticlassMCCLoss,
-    MVELoss,
-    Wasserstein,
-)
 from .message_passing import (
     AtomMessagePassing,
     BondMessagePassing,
@@ -29,18 +14,31 @@ from .message_passing import (
 )
 from .metrics import (
     MAE,
+    MSE,
     RMSE,
+    SID,
+    BCELoss,
     BinaryAccuracy,
     BinaryAUPRC,
     BinaryAUROC,
+    BinaryDirichletLoss,
     BinaryF1Score,
+    BinaryMCCLoss,
     BinaryMCCMetric,
     BoundedMAE,
     BoundedMSE,
     BoundedRMSE,
+    ChempropMetric,
+    CrossEntropyLoss,
+    EvidentialLoss,
+    LossFunctionRegistry,
     MetricRegistry,
+    MulticlassDirichletLoss,
+    MulticlassMCCLoss,
     MulticlassMCCMetric,
+    MVELoss,
     R2Score,
+    Wasserstein,
 )
 from .predictors import (
     BinaryClassificationFFN,
@@ -65,21 +63,15 @@ __all__ = [
     "SumAggregation",
     "NormAggregation",
     "AttentiveAggregation",
-    "MSE",
-    "SID",
-    "BCELoss",
-    "BinaryDirichletLoss",
-    "BinaryMCCLoss",
     "ChempropMetric",
-    "CrossEntropyLoss",
-    "EvidentialLoss",
     "LossFunctionRegistry",
-    "MulticlassDirichletLoss",
-    "MulticlassMCCLoss",
-    "MVELoss",
-    "Wasserstein",
+    "MetricRegistry",
+    "MSE",
     "MAE",
     "RMSE",
+    "BoundedMSE",
+    "BoundedMAE",
+    "BoundedRMSE",
     "BinaryAccuracy",
     "BinaryAUPRC",
     "BinaryAUROC",
@@ -91,6 +83,22 @@ __all__ = [
     "MetricRegistry",
     "MulticlassMCCMetric",
     "R2Score",
+    "MVELoss",
+    "EvidentialLoss",
+    "BCELoss",
+    "CrossEntropyLoss",
+    "BinaryMCCLoss",
+    "BinaryMCCMetric",
+    "MulticlassMCCLoss",
+    "MulticlassMCCMetric",
+    "BinaryAUROC",
+    "BinaryAUPRC",
+    "BinaryAccuracy",
+    "BinaryF1Score",
+    "BinaryDirichletLoss",
+    "MulticlassDirichletLoss",
+    "SID",
+    "Wasserstein",
     "MessagePassing",
     "AtomMessagePassing",
     "BondMessagePassing",
