+++ chemprop/utils/v1_to_v2.py
@@ -5,9 +5,8 @@ from lightning.pytorch.utilities.parsing import AttributeDict
 import torch
 
 from chemprop.nn.agg import AggregationRegistry
-from chemprop.nn.loss import LossFunctionRegistry
 from chemprop.nn.message_passing import AtomMessagePassing, BondMessagePassing
-from chemprop.nn.metrics import MetricRegistry
+from chemprop.nn.metrics import LossFunctionRegistry, MetricRegistry
 from chemprop.nn.predictors import PredictorRegistry
 from chemprop.nn.transforms import UnscaleTransform
 from chemprop.utils import Factory
